requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Enhanced algorithm dependencies
sentence-transformers>=2.2.0
numpy>=1.21.0

# Optional dependencies for better performance
pandas>=1.5.0